refactor(account): split code into modules and feature-gated sub-mods
